+++ manager-rs/dauth-service/src/rpc/handlers/backup_network.rs
@@ -1,153 +1,20 @@
 use std::sync::Arc;
 
-use auth_vector::types::ResStar;
-
 use crate::data::context::DauthContext;
 use crate::data::error::DauthError;
 use crate::data::signing::{self, SignPayloadType};
 use crate::data::vector::{AuthVectorReq, AuthVectorRes};
 use crate::manager;
+use crate::rpc::handlers::handler::DauthHandler;
 use crate::rpc::dauth::common::{AuthVector5G, UserIdKind};
-use crate::rpc::dauth::local::aka_confirm_resp;
-use crate::rpc::dauth::local::local_authentication_server::LocalAuthentication;
-use crate::rpc::dauth::local::{AkaConfirmReq, AkaConfirmResp, AkaVectorReq, AkaVectorResp};
-use crate::rpc::dauth::remote::{
-    backup_network_server::BackupNetwork, home_network_server::HomeNetwork,
-};
+use crate::rpc::dauth::remote::backup_network_server::BackupNetwork;
 use crate::rpc::dauth::remote::{delegated_auth_vector5_g, delegated_confirmation_share};
 use crate::rpc::dauth::remote::{
     DelegatedAuthVector5G, DelegatedConfirmationShare, EnrollBackupCommitReq,
     EnrollBackupCommitResp, EnrollBackupPrepareReq, EnrollBackupPrepareResp, FloodVectorReq,
-    FloodVectorResp, GetBackupAuthVectorReq, GetBackupAuthVectorResp, GetHomeAuthVectorReq,
-    GetHomeAuthVectorResp, GetHomeConfirmKeyReq, GetHomeConfirmKeyResp, GetKeyShareReq,
+    FloodVectorResp, GetBackupAuthVectorReq, GetBackupAuthVectorResp, GetKeyShareReq,
     GetKeyShareResp, WithdrawBackupReq, WithdrawBackupResp, WithdrawSharesReq, WithdrawSharesResp,
 };
-
-/// Handles all RPC calls to the dAuth service.
-pub struct DauthHandler {
-    pub context: Arc<DauthContext>,
-}
-
-#[tonic::async_trait]
-impl LocalAuthentication for DauthHandler {
-    /// Local request for a vector that will be used on this network.
-    /// No authentication is done.
-    async fn get_auth_vector(
-        &self,
-        request: tonic::Request<AkaVectorReq>,
-    ) -> Result<tonic::Response<AkaVectorResp>, tonic::Status> {
-        tracing::info!("Request: {:?}", request);
-
-        let av_request: AuthVectorReq;
-        match AuthVectorReq::from_req(request.into_inner()) {
-            Ok(req) => av_request = req,
-            Err(e) => return Err(tonic::Status::new(tonic::Code::Aborted, e.to_string())),
-        }
-
-        match manager::find_vector(self.context.clone(), &av_request).await {
-            Ok(av_result) => {
-                tracing::info!("Returning result: {:?}", av_result);
-                Ok(tonic::Response::new(av_result.to_resp()))
-            }
-            Err(e) => {
-                tracing::error!("Error while handling request for {:?}: {}", av_request, e);
-                Err(tonic::Status::new(tonic::Code::Aborted, e.to_string()))
-            }
-        }
-    }
-
-    /// Local request for to complete auth process for a vector.
-    /// No authentication is done.
-    async fn confirm_auth(
-        &self,
-        request: tonic::Request<AkaConfirmReq>,
-    ) -> Result<tonic::Response<AkaConfirmResp>, tonic::Status> {
-        tracing::info!("Request: {:?}", request);
-
-        let res_star = request.into_inner().res_star;
-        let res_star: auth_vector::types::ResStar =
-            res_star.try_into().or_else(|_e: Vec<u8>| {
-                Err(tonic::Status::new(
-                    tonic::Code::OutOfRange,
-                    "Unable to parse res_star",
-                ))
-            })?;
-
-        let kseaf = manager::confirm_auth_vector(self.context.clone(), res_star)
-            .await
-            .or_else(|e| Err(tonic::Status::new(tonic::Code::NotFound, e.to_string())))?;
-
-        let response_payload = AkaConfirmResp {
-            error: aka_confirm_resp::ErrorKind::NoError as i32,
-            kseaf: kseaf.to_vec(),
-        };
-
-        Ok(tonic::Response::new(response_payload))
-    }
-}
-
-#[tonic::async_trait]
-impl HomeNetwork for DauthHandler {
-    /// Remote request for a vector that will be generated on this network.
-    /// Checks for proper authentication and reputation.
-    async fn get_auth_vector(
-        &self,
-        request: tonic::Request<GetHomeAuthVectorReq>,
-    ) -> Result<tonic::Response<GetHomeAuthVectorResp>, tonic::Status> {
-        tracing::info!("Request: {:?}", request);
-
-        let message = request
-            .into_inner()
-            .message
-            .ok_or_else(|| tonic::Status::new(tonic::Code::NotFound, "No message received"))?;
-
-        let verify_result =
-            signing::verify_message(self.context.clone(), &message).or_else(|e| {
-                Err(tonic::Status::new(
-                    tonic::Code::Unauthenticated,
-                    format!("Failed to verify message: {}", e),
-                ))
-            })?;
-
-        match DauthHandler::get_home_auth_vector_hlp(self.context.clone(), verify_result).await {
-            Ok(result) => Ok(result),
-            Err(e) => Err(tonic::Status::new(
-                tonic::Code::Aborted,
-                format!("Error while handling request: {}", e),
-            )),
-        }
-    }
-
-    /// Remote request for to complete auth process for a vector.
-    /// Checks for proper authentication.
-    async fn get_confirm_key(
-        &self,
-        request: tonic::Request<GetHomeConfirmKeyReq>,
-    ) -> Result<tonic::Response<GetHomeConfirmKeyResp>, tonic::Status> {
-        tracing::info!("Request: {:?}", request);
-
-        let message = request
-            .into_inner()
-            .message
-            .ok_or_else(|| tonic::Status::new(tonic::Code::NotFound, "No message received"))?;
-
-        let verify_result =
-            signing::verify_message(self.context.clone(), &message).or_else(|e| {
-                Err(tonic::Status::new(
-                    tonic::Code::Unauthenticated,
-                    format!("Failed to verify message: {}", e),
-                ))
-            })?;
-
-        match DauthHandler::get_confirm_key_hlp(self.context.clone(), verify_result).await {
-            Ok(result) => Ok(result),
-            Err(e) => Err(tonic::Status::new(
-                tonic::Code::Aborted,
-                format!("Error while handling request: {}", e),
-            )),
-        }
-    }
-}
 
 #[tonic::async_trait]
 impl BackupNetwork for DauthHandler {
@@ -406,68 +273,6 @@ impl DauthHandler {
     }
 
     /* Specific helpers */
-    async fn get_home_auth_vector_hlp(
-        context: Arc<DauthContext>,
-        verify_result: SignPayloadType,
-    ) -> Result<tonic::Response<GetHomeAuthVectorResp>, DauthError> {
-        if let SignPayloadType::GetHomeAuthVectorReq(payload) = verify_result {
-            let user_id = std::str::from_utf8(payload.user_id.as_slice())?.to_string();
-
-            // TODO: Handle reputation
-
-            let av_result = manager::generate_auth_vector(
-                context.clone(),
-                &AuthVectorReq {
-                    user_id: user_id.to_string(),
-                },
-            )
-            .await?;
-
-            let payload = delegated_auth_vector5_g::Payload {
-                serving_network_id: context.local_context.id.clone(),
-                v: Some(AuthVector5G {
-                    rand: av_result.rand.to_vec(),
-                    xres_star_hash: av_result.xres_star_hash.to_vec(),
-                    autn: av_result.autn.to_vec(),
-                    seqnum: av_result.seqnum,
-                }),
-            };
-
-            Ok(tonic::Response::new(GetHomeAuthVectorResp {
-                vector: Some(DelegatedAuthVector5G {
-                    message: Some(signing::sign_message(
-                        context,
-                        signing::SignPayloadType::DelegatedAuthVector5G(payload),
-                    )),
-                }),
-            }))
-        } else {
-            Err(DauthError::InvalidMessageError(format!(
-                "Incorrect message type: {:?}",
-                verify_result
-            )))
-        }
-    }
-
-    async fn get_confirm_key_hlp(
-        context: Arc<DauthContext>,
-        verify_result: SignPayloadType,
-    ) -> Result<tonic::Response<GetHomeConfirmKeyResp>, DauthError> {
-        if let SignPayloadType::GetHomeConfirmKeyReq(payload) = verify_result {
-            let res_star: ResStar = payload.res_star.as_slice().try_into()?;
-
-            let kseaf = manager::confirm_auth_vector(context, res_star).await?;
-
-            Ok(tonic::Response::new(GetHomeConfirmKeyResp {
-                kseaf: kseaf.to_vec(),
-            }))
-        } else {
-            Err(DauthError::InvalidMessageError(format!(
-                "Incorrect message type: {:?}",
-                verify_result
-            )))
-        }
-    }
 
     async fn enroll_backup_prepare_hlp(
         context: Arc<DauthContext>,
