+++ manager-rs/dauth-service/src/rpc/handlers/mod.rs
@@ -1,0 +1,4 @@
+pub mod backup_network;
+pub mod home_network;
+pub mod local_authentication;
+pub mod handler;
